+++ Changes.md
@@ -1,3 +1,13 @@
+## v1.1.2 (July 6, 2020)
+
+### WickedWhims Support
+- Cycle now has basic support for the mod "WickedWhims". I don't understand why some of you like this mod. I've added support because many people seem to want it and some are just ignoring my warnings about the two mods being incompatible.
+- "Woohoo" interactions added by WickedWhims now add sperm to sims.
+- WickedWhims's period system is now disabled and replaced by Cycle.
+- Cycle's condoms boxes must be unpacked before they can be used with WickedWhims's interactions.
+
+______________________________
+
 ## v1.1.1 (July 2, 2020)
 
 ### Performance Improvements
@@ -9,6 +19,8 @@
 ### Other Improvements
 - Added system to try to always remember which sim is which, even if another mod goes and changes their sim id.
 
+______________________________
+
 ## v1.1.0 (June 24, 2020)
 
 ### Fixed bugs
@@ -17,6 +29,8 @@
 - Pregnant teens now receive the pregnancy moodlets and correctly physically show their pregnancy.
 - Reproductive systems and dot information are now correctly load data when loading a game from the main menu.
 - The dot cycle tracker app can now exit out of pregnancy mode.
+
+______________________________
 
 ## v1.0.0 (June 19, 2020)
  - Initial release